backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3

